requirements.txt
textual>=0.41.0
rich>=13.0.0
pyfiglet>=0.8
paramiko==3.4.0
pywinrm==0.4.3
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
# Optional: for SHAP explainability (pip install shap)
# shap>=0.44.0
# Optional: for SMOTE class balancing in training (pip install imbalanced-learn)
# imbalanced-learn>=0.11.0
